Icon apifixes

+++ app/src/main/java/com/leandrososa/weatherapp/router/IRouter.kt
@@ -7,7 +7,7 @@ interface IRouter {
 sealed class Route(val id: String) {
     data object Cities: Route("cities")
     data class Weather(
-        var lat: Double = 0.0,
-        var lon: Double = 0.0
+        var lat: Float = 0.0f,
+        var lon: Float = 0.0f
     ): Route("weather")
 }
